Separate grouped display logic from ResultSetPresenter
Grouped display logic is customised to the requirements of one finder
at the moment, so move the grouping logic to a bespoke presenter which
is only used when the specific ordering logic is activated.
