+++ app/views/finders/_results.mustache
@@ -7,9 +7,9 @@
 
 <ul class="finder-results" data-module="track-click">
   {{#display_grouped_results }}
-    {{#documents_by_facets}}
+    {{#grouped_documents}}
       {{> finders/_grouped_results }}
-    {{/documents_by_facets}}
+    {{/grouped_documents}}
   {{/display_grouped_results}}
 
   {{^display_grouped_results}}
